Clear community-review Warning and Recommendation findings
Align the plugin with eslint-plugin-obsidianmd recommended without
bumping 1.0.2 or shipping a bundle:

- Popout-safe window timers, Document.createEl, sentence-case copy
- Node http desktop MCP fetch (R-012); drop node-fetch
- Bundle under 5MB via LangSmith stub and MCP-nested ajv dedupe
- App localStorage for native runtime paths; local debounce/deepEqual
- getSettingDefinitions() plus display() for minAppVersion 1.11.4
- Untyped system_fingerprint reads and legacy SSE construction
- Reviewer warnings.config.mjs; prismjs 1.30.0 override

main.js and meta.json remain uncommitted.

+++ src/utils/llm/sse.ts
@@ -36,8 +36,7 @@ export async function* parseJsonSseStream<T>(
   }
 
   for await (const chunk of body) {
-    const chunkText =
-      typeof chunk === 'string' ? chunk : decoder.decode(chunk as Uint8Array)
+    const chunkText = typeof chunk === 'string' ? chunk : decoder.decode(chunk)
     buffer += chunkText
     yield* flushSseBuffer(buffer, (next) => {
       buffer = next
